Remove preserveEmptyContent setting from experimental formatter
